requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
sentence-transformers
faiss-cpu
numpy
requests
streamlit
python-multipart